# docs/swagger.yaml
basePath: /
definitions:
...
  email-service_internal_model.Recipient:
    description: Stores information about a person who can receive emails, including
      custom data.
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      status:
        type: string
      updated_at:
        type: string
    type: object
  email-service_internal_model.RecipientGroup:
    description: Defines a group of recipients, which can be a fixed list (static)
      or rule-based (dynamic).
    properties:
      created_at:
        type: string
      created_by_user:
        $ref: '#/definitions/email-service_internal_model.User'
      created_by_user_id:
        type: integer
      description:
        type: string
      group_type:
        description: '''static'' or ''dynamic'''
        type: string
      id:
        type: integer
      members:
        items:
          $ref: '#/definitions/email-service_internal_model.Recipient'
        type: array
      name:
        type: string
      rules:
        items:
          $ref: '#/definitions/email-service_internal_model.RecipientGroupRule'
        type: array
      updated_at:
        type: string
    type: object
  email-service_internal_model.RecipientGroupRule:
    description: A rule used to dynamically include recipients in a group based on
      their attributes.
    properties:
      field:
        description: e.g., "status", "metadata.country"
        type: string
      group_id:
        type: integer
      id:
        type: integer
      operator:
        description: e.g., "equals", "contains"
        type: string
      value:
        type: string
    type: object
...
  email-service_internal_model.User:
    description: Defines a user of this email service platform, with roles and access
      rights.
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_admin:
        type: boolean
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  internal_handler.CreateGroupRequest:
    properties:
      description:
        type: string
      group_type:
        enum:
        - static
        - dynamic
        type: string
      member_ids:
        items:
          type: integer
        type: array
      name:
        type: string
      rules:
        items:
          $ref: '#/definitions/email-service_internal_model.RecipientGroupRule'
        type: array
    required:
    - group_type
    - name
    type: object
...
  internal_handler.ModifyMembersRequest:
    properties:
      member_ids:
        items:
          type: integer
        type: array
    required:
    - member_ids
    type: object
  internal_handler.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/internal_handler.Pagination'
    type: object
  internal_handler.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  internal_handler.Response:
    properties:
      data: {}
      error:
        type: string
    type: object
...
  internal_handler.UpdateGroupRequest:
    properties:
      description:
        type: string
      name:
        type: string
      rules:
        items:
          $ref: '#/definitions/email-service_internal_model.RecipientGroupRule'
        type: array
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: email-service/internal/handler
    url: http://www.swagger.io/support
  description: This is a powerful email service API for batch sending and management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Email Service API
  version: "1.0"
paths:
...
  /api/v1/recipient-groups:
    get:
      description: Retrieves a paginated list of all recipient groups.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of groups
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/email-service_internal_model.RecipientGroup'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: List recipient groups
      tags:
      - Recipient Groups
    post:
      consumes:
      - application/json
      description: Creates either a static or a dynamic recipient group. For static
        groups, provide member_ids. For dynamic groups, provide rules.
      parameters:
      - description: Group Creation Request
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created group
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_model.RecipientGroup'
              type: object
        "400":
          description: Invalid request body or parameters
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a recipient group
      tags:
      - Recipient Groups
  /api/v1/recipient-groups/{id}:
    delete:
      description: Deletes a recipient group and its associations.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Successfully deleted group
        "400":
          description: Invalid group ID
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete a recipient group
      tags:
      - Recipient Groups
    get:
      description: Retrieves details of a specific recipient group by its ID.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Group details
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_model.RecipientGroup'
              type: object
        "400":
          description: Invalid request body or parameters
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a recipient group
      tags:
      - Recipient Groups
    put:
      consumes:
      - application/json
      description: Updates a recipient group's details. For dynamic groups, rules
        can be updated.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Group Update Request
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated group
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_model.RecipientGroup'
              type: object
        "400":
          description: Invalid request body or parameters
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Update a recipient group
      tags:
      - Recipient Groups
  /api/v1/recipient-groups/{id}/members:
    delete:
      consumes:
      - application/json
      description: Removes a list of recipients from a static group.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Members to Remove
        in: body
        name: members
        required: true
        schema:
          $ref: '#/definitions/internal_handler.ModifyMembersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully removed members
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Invalid request body or parameters
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Remove members from a static group
      tags:
      - Recipient Groups
    post:
      consumes:
      - application/json
      description: Adds a list of recipients to a static group.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Members to Add
        in: body
        name: members
        required: true
        schema:
          $ref: '#/definitions/internal_handler.ModifyMembersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully added members
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Invalid request body or parameters
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Add members to a static group
      tags:
      - Recipient Groups
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"